Exit if nanoparticle is too large
If a large number (e.g. millions) of particles are required to
construct a nanoparticle (in the case of a small sigma and large
nanoparticle radius) the build script requires too much memory and
can lock up the computer. I've added a check to help catch this
that can be manually overridden if desired.