Corrected bug in NP creation when `bvf` specified
Previously when `bvf` was specified a check was performed to
make sure three beads did not intersect, as the volume of the
tri-bead intersection cannot be easily computed so the bead
volume fraction cannot be computed. However, this check was
incorrect. This commit addresses this by now checking for tri-bead
intersections through trilateration.

Additionally, calculation of `bvf` was previously performed
incorrectly. The volume of the intersections was twice the actual
value due to duplicated distances in the computed distance array.
This is now addressed by dividing that volume by two.