Update NanoOpt for recent mBuild changes
Currently, `NanoOpt` has not been updated to support the major box and
periodicity overhaul that occured with newer versions of mBuild.

This PR employs those changes, while also updating the README so users
who want to use the version that was used during the publication should
be able to do so as well. They just need to use a version at a specific
git commit, which has been added to the readme.